+++ mido/midifiles.py
@@ -124,9 +124,6 @@ class ByteWriter(object):
 
 
 class MidiTrack(list):
-    def __init__(self):
-        list.__init__([])
-
     @property
     def name(self):
         """Name of the track.
@@ -154,6 +151,14 @@ class MidiTrack(list):
         else:
             # No track name found, add one.
             self.insert(0, MetaMessage('track_name', name=name, time=0))
+
+    def __getitem__(self, index_or_slice):
+        # Call slice on our parent class (a list).
+        lst = super().__getitem__(index_or_slice)
+        # The slice returns a list. Cast the return object to the same
+        # class we are.
+        # Todo: this make a copy of the list. Is there a better way?
+        return self.__class__(lst)
 
     def __repr__(self):
         return '<midi track {!r} {} messages>'.format(self.name, len(self))