+++ mido/test_midifiles.py
@@ -1,4 +1,5 @@
 from .midifiles_meta import tempo2bpm, bpm2tempo
+from .midifiles import MidiTrack
 
 def test_tempo2bpm_bpm2tempo():
     for bpm, tempo in [
@@ -9,3 +10,10 @@ def test_tempo2bpm_bpm2tempo():
     ]:
         assert bpm == tempo2bpm(tempo)
         assert tempo == bpm2tempo(bpm)
+
+
+def test_track_slice():
+    track = MidiTrack()
+
+    # Slice should return MidiTrack object.
+    assert isinstance(track[::], MidiTrack)